refactor: organize internal modules
