+++ src/abstractions/app-configuration-renderer.ts
