+++ src/abstractions/application/AppSessionAuthData.ts
@@ -1,0 +1,8 @@
+import { AppSessionData } from "./AppSessionData";
+
+type AuthData = {
+  jwtToken: string;
+  dopplerAccountName: string;
+};
+
+export type AppSessionAuthData = AppSessionData<AuthData>;
