+++ src/abstractions/application/AppSessionData.ts
@@ -1,0 +1,6 @@
+export type AppSessionData<T> =
+  | { readonly status: "unknown" }
+  | { readonly status: "non-authenticated" }
+  | ({
+      readonly status: "authenticated";
+    } & T);
