+++ src/abstractions/application/AppSessionStateMonitor.ts
@@ -1,0 +1,4 @@
+export interface AppSessionStateMonitor {
+  readonly start: () => void;
+  onSessionUpdate: () => void;
+}
