+++ src/abstractions/html-editor-api-client.ts
@@ -1,5 +1,5 @@
-import { Result } from "../common/result-types";
-import { Content } from "../domain/content";
+import { Result } from "./result-types";
+import { Content } from "./domain/content";
 
 export interface HtmlEditorApiClient {
   getCampaignContent: (campaignId: string) => Promise<Result<Content>>;
