--- src/abstractions/index.ts
@@ -1,2 +1,0 @@
-export type { AppConfiguration } from "./configuration";
-export type { AppServices } from "./services";
