+++ src/abstractions/result-types.ts
