--- src/components/LoadingScreen.tsx
@@ -1,7 +1,0 @@
-export const LoadingScreen = () => {
-  return (
-    <div className="wrapper-loading" style={{ height: "99vh" }}>
-      <div className="loading-page" />
-    </div>
-  );
-};
