+++ src/components/application/AppSessionStateContext.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { AppSessionUserData } from "../../abstractions/application";
+
+export const AppSessionStateContext = createContext<AppSessionUserData>({
+  status: "unknown",
+});
