+++ src/components/application/LoadingScreen.tsx
@@ -1,0 +1,11 @@
+export const LoadingScreen = () => {
+  return (
+    <div
+      className="wrapper-loading"
+      style={{ height: "99vh" }}
+      data-testid="loading-page"
+    >
+      <div className="loading-page" />
+    </div>
+  );
+};
