+++ src/components/application/useAppServices.ts
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { AppServicesContext } from "./AppServicesContext";
+
+export const useAppServices = () => useContext(AppServicesContext);
