+++ src/components/application/useAppSessionUserData.ts
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { AppSessionStateContext } from "./AppSessionStateContext";
+
+export const useAppSessionUserData = () => useContext(AppSessionStateContext);
