+++ src/implementations/session-mfe/SessionMfeAppSessionStateClient.ts
@@ -3,27 +3,8 @@ import {
   AppSessionUserData,
   AppSessionStateAccessor,
   AppSessionStateMonitor,
-} from "../../abstractions/app-session";
-
-export const DOPPLER_SESSION_STATE_UPDATE_EVENT_TYPE =
-  "doppler-session-state-update";
-
-export type DopplerSessionState =
-  | undefined
-  | { status: "non-authenticated" }
-  | {
-      status: "authenticated";
-      jwtToken: string;
-      dopplerAccountName: string;
-      lang: string;
-      rawDopplerUserData: any;
-    };
-
-declare global {
-  interface Window {
-    dopplerSessionState: DopplerSessionState;
-  }
-}
+} from "../../abstractions/application";
+import { DOPPLER_SESSION_STATE_UPDATE_EVENT_TYPE } from "./session-mfe-conventions";
 
 export class SessionMfeAppSessionStateClient
   implements AppSessionStateMonitor, AppSessionStateAccessor
